requirements.txt
numpy==1.13.3
python_speech_features
tensorflow==1.4.0
PyYAML
tqdm
librosa
scikit-learn